Updated biome generation to be per subchunk + introduced concept of probability multipliers

+++ DrifterEngine/src/Random/NoiseLayer.cpp
@@ -12,7 +12,7 @@ void drft::rng::NoiseLayer::createFromJSON(const rapidjson::Value& json)
 {
     if (json.HasMember("Resolution"))
     {
-        _resolution = json["Resolution"].GetInt();
+        _resolution = json["Resolution"].GetFloat();
     }
     if (json.HasMember("Octaves"))
     {
